// apps/creator-hub/modules/matchmaking/enums/DefaultConfigurationSignals.ts
enum DefaultConfigurationSignals {
  Occupancy = 'Occupancy',
  AgeDifference = 'AgeDifference',
  Language = 'Language',
  Latency = 'Latency',
  PreferredPlayers = 'PreferredPlayers',
  DeviceType = 'DeviceType',
  VoiceChat = 'VoiceChat',
  PlayHistory = 'PlayHistory',
  TextChat = 'TextChat',
}

export default DefaultConfigurationSignals;